Create reusable components for game modal stats display
- All football stat displays are now reusable, including PBP
- I have also created these for hockey, however I have not implemented
them into the main file as I am unable to test them with no games data
available at present
